Server/requirements.txt
fastapi==0.115.11
motor==3.7.0
numpy==2.2.3
pandas==2.2.3
pydantic==2.10.6
python-dotenv==1.0.1
rdkit_pypi==2022.9.5
starlette==0.46.1
torch==2.6.0
transformers==4.49.0
uvicorn==0.34.0
